perf(votes): keep scrutin detail ISR by reading the type tab client-side
The [slug] route awaited searchParams (date-archive type tab), which opted the
whole route into dynamic rendering and silently disabled ISR — every detail
request re-ran the heavy 577-deputy query. Move the type filter to a client
component (useSearchParams) so the route stays ISR-cacheable; date archives and
?type= still work.
